+++ types/check_test.go
@@ -112,9 +112,9 @@ func TestProxyRequestsValidate(t *testing.T) {
 	p.SplayCoverage = 90
 
 	// Invalid entity attributes
-	p.EntityAttributes = []string{`entity.Class = "proxy"`}
+	p.EntityAttributes = []string{`entity.EntityClass = "proxy"`}
 	assert.Error(t, p.Validate())
-	p.EntityAttributes = []string{`entity.Class == "proxy"`}
+	p.EntityAttributes = []string{`entity.EntityClass == "proxy"`}
 
 	// Valid proxy request
 	assert.NoError(t, p.Validate())